pyproject.toml
[project]
name = "docs-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastmcp>=2.12.2",
    "httpx>=0.28.1",
]
